pyproject.toml
[project]
name = "inventory-management"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "agents",
    "openai",
    "openai-agents>=0.2.8",
    "pydantic>=2",
    "python-dotenv",
]
[project.scripts]
inventory-agents = "inventory_management.main:start"
